testing updated. New tests created and duplicate removed
  27 tests passed

=============================== Coverage summary ===============================
Statements   : 100% ( 195/195 )
Branches     : 100% ( 35/35 )
Functions    : 100% ( 22/22 )
Lines        : 100% ( 195/195 )
================================================================================
